+++ src/promptflow/promptflow/_core/run_tracker.py
@@ -15,7 +15,7 @@ from promptflow._utils.dataclass_serializer import serialize
 from promptflow._utils.exception_utils import ExceptionPresenter
 from promptflow._utils.logger_utils import flow_logger
 from promptflow._utils.openai_metrics_calculator import OpenAIMetricsCalculator
-from promptflow.contracts.multimedia import Image
+from promptflow.contracts.multimedia import PFBytes
 from promptflow.contracts.run_info import FlowRunInfo, RunInfo, Status
 from promptflow.contracts.run_mode import RunMode
 from promptflow.contracts.tool import ConnectionType
@@ -240,11 +240,8 @@ class RunTracker(ThreadLocalSingleton):
             return ConnectionType.serialize_conn(val)
         if self.allow_generator_types and isinstance(val, GeneratorType):
             return str(val)
-        if isinstance(val, Image):
-            # Images will be persisted when persisting flow run or node run, so no need to persist it here.
-            return val
         try:
-            json.dumps(val)
+            json.dumps(val, default=PFBytes.default_json_encoder)
             return val
         except Exception:
             if not warning_msg: