+++ src/promptflow/tests/sdk_cli_azure_test/e2etests/test_run_operations.py
@@ -259,6 +259,51 @@ class TestFlowRun:
         # error info will store in run dict
         assert "error" in run._to_dict()
 
+    def test_archive_and_restore_run(self, remote_client):
+        from promptflow._sdk._constants import RunHistoryKeys
+
+        run_meta_data = RunHistoryKeys.RunMetaData
+        hidden = RunHistoryKeys.HIDDEN
+
+        run_id = "4cf2d5e9-c78f-4ab8-a3ee-57675f92fb74"
+
+        # test archive
+        remote_client.runs.archive(run=run_id)
+        run_data = remote_client.runs._get_run_from_run_history(run_id, original_form=True)[run_meta_data]
+        assert run_data[hidden] is True
+
+        # test restore
+        remote_client.runs.restore(run=run_id)
+        run_data = remote_client.runs._get_run_from_run_history(run_id, original_form=True)[run_meta_data]
+        assert run_data[hidden] is False
+
+    def test_update_run(self, remote_client):
+        run_id = "4cf2d5e9-c78f-4ab8-a3ee-57675f92fb74"
+        test_mark = str(uuid.uuid4())
+
+        new_display_name = f"test_display_name_{test_mark}"
+        new_description = f"test_description_{test_mark}"
+        new_tags = {"test_tag": test_mark}
+
+        run = remote_client.runs.update(
+            run=run_id,
+            display_name=new_display_name,
+            description=new_description,
+            tags=new_tags,
+        )
+        assert run.display_name == new_display_name
+        assert run.description == new_description
+        assert run.tags["test_tag"] == test_mark
+
+        # test wrong type of parameters won't raise error, just log warnings and got ignored
+        run = remote_client.runs.update(
+            run=run_id,
+            tags={"test_tag": {"a": 1}},
+        )
+        assert run.display_name == new_display_name
+        assert run.description == new_description
+        assert run.tags["test_tag"] == test_mark
+
     def test_run_with_additional_includes(self, remote_client, pf, runtime):
         run = pf.run(
             flow=f"{FLOWS_DIR}/web_classification_with_additional_include",