+++ src/promptflow/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features Added
 - [Executor] Add average execution time and estimated execution time to batch run logs
+- [SDK/CLI] Support `pfazure run archive/restore/update`.
 - **Telemetry**: enable telemetry and won't collect by default, use `pf config set cli.telemetry_enabled=true` to opt in.
 
 ### Bugs Fixed